services/ingestion/pyproject.toml
[tool.poetry]
name = "rag-ingestion-service"
version = "0.1.0"
description = "Data Ingestion & Vectorization Service — file parsing, chunking, embedding, Qdrant storage"
authors = ["RAG Team"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
rag-shared = {path = "../../shared", develop = true}
numpy = ">=2.0"
fastapi = "^0.128"
uvicorn = {extras = ["standard"], version = "^0.40"}
python-multipart = "^0.0.20"
llama-index-core = "^0.13"
llama-index-embeddings-dashscope = ">=0.2"
llama-index-vector-stores-qdrant = "^0.9"
llama-index-readers-file = ">=0.2"
qdrant-client = "^1.16"
jieba = "^0.42"
docx2txt = "^0.8"
langchain-text-splitters = ">=0.2"
pypdf = "^5.0"
python-dotenv = "^1.0"
pydantic-settings = "^2.0"
pymupdf4llm = "^0.3.4"
llama-index-embeddings-openai = "^0.5.1"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
